+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2024-10-05:
+  - updated packages to include improvements in the `Corpora` library. These improvements now allow specifying XPath expressions returning a single value rather than a nodeset as XML structure parser value template arguments.
+  - set XML sentence parser result value to be the count of the (whitespace-normalized) characters in the sentence.
 - 2024-10-02: updated packages (including Corpora upgraded to .NET 8).
 
 ## [4.1.0] - 2024-09-19